split reading and writing tests
