+++ src/tests/testread.cpp
@@ -1,0 +1,342 @@
+/*
+ *  testread.cpp
+ *  LibKEA
+ *
+ *  Created by Pete Bunting on 02/07/2012.
+ *  Copyright 2012 LibKEA. All rights reserved.
+ *
+ *  This file is part of LibKEA.
+ *
+ *  Permission is hereby granted, free of charge, to any person 
+ *  obtaining a copy of this software and associated documentation 
+ *  files (the "Software"), to deal in the Software without restriction, 
+ *  including without limitation the rights to use, copy, modify, 
+ *  merge, publish, distribute, sublicense, and/or sell copies of the 
+ *  Software, and to permit persons to whom the Software is furnished 
+ *  to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be 
+ *  included in all copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, 
+ *  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES 
+ *  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. 
+ *  IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR 
+ *  ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF 
+ *  CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION 
+ *  WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
+ */
+ 
+#include <stdio.h>
+#include <stdlib.h>
+#include <iostream>
+#include "libkea/KEAImageIO.h"
+#include "testsupport.h"
+
+#define OV_XSIZE 300
+#define OV_YSIZE 350
+
+int main()
+{
+    try
+    {
+        kealib::KEAImageIO io;
+
+        std::string test_kea_file = "test_uint8_t.kea";
+        
+        std::cout << "Opening file" << std::endl;
+        HighFive::File *h5file = kealib::KEAImageIO::openKeaH5RDOnly(test_kea_file);
+        io.openKEAImageHeader(h5file);
+        std::cout << "Opened file" << std::endl;
+        
+        auto spatialInfo2 = getSpatialInfo(10);
+        auto readinfo2 = io.getSpatialInfo();
+        if( !compareSpatialInfo(&spatialInfo2, readinfo2))
+        {
+            return 1;
+        }
+
+        std::cout << "Reading some image data" << std::endl;
+        uint8_t *pReadData = (uint8_t*)calloc(readinfo2->xSize * readinfo2->ySize, sizeof(uint8_t));
+        io.readImageBlock2Band(1, pReadData, 0, 0, readinfo2->xSize, readinfo2->ySize, readinfo2->xSize, readinfo2->ySize, kealib::kea_8uint);
+        std::cout << "Read some image data" << std::endl;
+
+        uint8_t *pCheckData = createDataForType<uint8_t>(readinfo2->xSize, readinfo2->ySize);
+        std::cout << "Comparing data written and read data" << std::endl;
+        if( !compareData<uint8_t>(pReadData, pCheckData, readinfo2->xSize, readinfo2->ySize))
+        {
+            return 1;
+        }
+        std::cout << "Data compared" << std::endl;
+        free(pReadData);
+        free(pCheckData);
+        
+        if( io.maskCreated(1) )
+        {
+            std::cout << "Mask created" << std::endl;
+        }
+        else
+        {
+            std::cout << "Mask NOT created" << std::endl;
+            return 1;
+        }
+        
+        uint8_t *pReadMask = (uint8_t*)calloc(readinfo2->xSize * readinfo2->ySize, sizeof(uint8_t));
+        io.readImageBlock2BandMask(1, pReadMask, 0, 0, readinfo2->xSize, readinfo2->ySize, readinfo2->xSize, readinfo2->ySize, kealib::kea_8uint);
+        std::cout << "Read some mask data" << std::endl;
+
+        uint8_t *pMaskData = createDataForType<uint8_t>(readinfo2->xSize, readinfo2->ySize);
+
+        std::cout << "Comparing data written and read mask" << std::endl;
+        if( !compareData<uint8_t>(pReadMask, pMaskData, readinfo2->xSize, readinfo2->ySize))
+        {
+            return 1;
+        }
+        std::cout << "Mask compared" << std::endl;
+
+        if( io.getImageMetaData("Test1") != "Value1" )
+        {
+            std::cout << "Error reading metadata" << std::endl;
+            return 1;
+        }
+        if( io.getImageMetaData("Test2") != "Value2" )
+        {
+            std::cout << "Error reading metadata" << std::endl;
+            return 1;
+        }
+        
+        auto names = io.getImageMetaDataNames();
+        if( names.size() != 4 )
+        {
+            std::cout << "Wrong number of image metadata items" << std::endl;
+            return 1;
+        }
+        if( std::find(names.begin(), names.end(), "Test1") == names.end())
+        {
+            std::cout << "Can't find Test1 in image metadata" << std::endl;
+            return 1;
+        }
+        if( std::find(names.begin(), names.end(), "Test2") == names.end())
+        {
+            std::cout << "Can't find Test2 in image metadata" << std::endl;
+            return 1;
+        }
+        
+        auto name_values = io.getImageMetaData();
+        if( name_values.size() != 4 )
+        {
+            std::cout << "Wrong number of image metadata items/values" << std::endl;
+            return 1;
+        }
+        for( auto itr = name_values.begin(); itr != name_values.end(); itr++)
+        {
+            auto name = (*itr).first;
+            if( (name != "Test1") && (name != "Test2") && (name != "Test98") && (name != "Test99"))
+            {
+                std::cout << "Unexpected name " << name << std::endl;
+                return 1;
+            }
+            auto value = (*itr).second;
+            if( ( value != "Value1") && (value != "Value2") && (value != "Value98") && (value != "Value99"))
+            {
+                std::cout << "Unexpected name " << value << std::endl;
+                return 1;
+            }
+        }
+        std::cout << "Successfully read dataset metadata" << std::endl;
+
+        if( io.getImageBandMetaData(1, "BandTest1") != "Value1" )
+        {
+            std::cout << "Error reading metadata" << std::endl;
+            return 1;
+        }
+        if( io.getImageBandMetaData(1, "BandTest2") != "Value2" )
+        {
+            std::cout << "Error reading metadata" << std::endl;
+            return 1;
+        }
+
+        auto bandnames = io.getImageBandMetaDataNames(1);
+        if( bandnames.size() != 4 )
+        {
+            std::cout << "Wrong number of band metadata items" << std::endl;
+            return 1;
+        }
+        if( std::find(bandnames.begin(), bandnames.end(), "BandTest1") == names.end())
+        {
+            std::cout << "Can't find BandTest1 in image metadata" << std::endl;
+            return 1;
+        }
+        if( std::find(bandnames.begin(), bandnames.end(), "BandTest2") == names.end())
+        {
+            std::cout << "Can't find BandTest2 in image metadata" << std::endl;
+            return 1;
+        }
+        
+        auto bandname_values = io.getImageBandMetaData(1);
+        if( bandname_values.size() != 4 )
+        {
+            std::cout << "Wrong number of image metadata items/values" << std::endl;
+            return 1;
+        }
+        for( auto itr = bandname_values.begin(); itr != bandname_values.end(); itr++)
+        {
+            auto name = (*itr).first;
+            if( (name != "BandTest1") && (name != "BandTest2") && (name != "BandTest98") && (name != "BandTest99"))
+            {
+                std::cout << "Unexpected name " << name << std::endl;
+                return 1;
+            }
+            auto value = (*itr).second;
+            if( ( value != "Value1") && (value != "Value2") && (value != "Value98") && (value != "Value99"))
+            {
+                std::cout << "Unexpected value " << value << std::endl;
+                return 1;
+            }
+        }
+
+        std::cout << "Successfully read band metadata" << std::endl;
+
+        auto desc1 = io.getImageBandDescription(1);
+        auto desc2 = io.getImageBandDescription(2);
+        if( (desc1 != "HBand1Desc") || (desc2 != "HBand2Desc"))
+        {
+            std::cout << "band descriptions did not match " << desc1 << " " << desc1 << std::endl;
+            return 1;
+        }
+        std::cout << "Read band descriptions ok" << std::endl;
+        
+        // read nodata back as different type
+        uint32_t u32nodata = 0;
+        io.getNoDataValue(1, &u32nodata, kealib::kea_32uint);
+        if( u32nodata != 999 )
+        {
+            std::cout << "Wrong nodata value read" << std::endl;
+            return 1;
+        }
+        // band 2 nodata is undefined
+        double dnodata = -1;
+        io.getNoDataValue(2, &dnodata, kealib::kea_64float);
+        if( dnodata != -1 )
+        {
+            std::cout << "Nodata not properly undefined" << std::endl;
+            return 1;
+        }
+        std::cout << "Read nodata ok" << std::endl;
+        
+        if( io.getGCPProjection() != "KmKmqw" )
+        {
+            std::cout << "GCP Projection not correctly read" << std::endl;
+            return 1;
+        }
+        
+        if( io.getGCPCount() != 2 )
+        {
+            std::cout << "Wrong number of GCPs read" << std::endl;
+            return 1;
+        }
+        
+        auto readgcps = io.getGCPs();
+        auto pGCPs = getGCPData();
+        if(!compareGCPs(pGCPs, readgcps))
+        {
+            std::cout << "GCPs don't match" << std::endl;
+            return 1;
+        }
+        freeGCPData(pGCPs);
+        freeGCPData(readgcps);
+        
+        if( io.getAttributeTableChunkSize(1) != kealib::KEA_ATT_CHUNK_SIZE )
+        {
+            std::cout << "Attribute chunk size does not match" << std::endl;
+            return 1;
+        }
+        
+        if( io.getNumOfImageBands() != 2)
+        {
+            std::cout << "Wrong number of image bands" << std::endl;
+            return 1; 
+        }
+        
+        if( io.getImageBlockSize(1) != 512 ) 
+        {
+            std::cout << "image block size does not match" << std::endl;
+            return 1;
+        }
+        
+        // band 2 should be the default (kea_continuous)
+        // but we changed band 1
+        if( (io.getImageBandLayerType(1) != kealib::kea_continuous) ||
+            (io.getImageBandLayerType(2) != kealib::kea_thematic) )
+        {
+            std::cout << "layer types do not match" << std::endl;
+            return 1;
+        }
+        
+        // band 2 should be the default (kea_generic) but 
+        // we changed band 1
+        if( ( io.getImageBandClrInterp(1) != kealib::kea_generic) ||
+            (io.getImageBandClrInterp(2) != kealib::kea_redband))
+        {
+            std::cout << "colour interp does not match" << std::endl;
+            return 1;
+        }
+
+        // overviews        
+        if( io.getNumOfOverviews(1) != 2 )
+        {
+            std::cout << "should be 2 overviews" << std::endl;
+            return 1;
+        }
+        if(io.getNumOfOverviews(2) != 0)
+        {
+            std::cout << "should be 0 overviews" << std::endl;
+            return 1;
+        }
+        if( io.getOverviewBlockSize(1, 0) != 300)
+        {
+            std::cout << "wrong overview block size" << std::endl;
+            return 1;
+        }
+        
+        uint64_t xsize, ysize;
+        io.getOverviewSize(1, 0, &xsize, &ysize);
+        if( (xsize != OV_XSIZE) || (ysize != OV_YSIZE))
+        {
+            std::cout << "wrong overview size" << std::endl;
+            return 1;
+        }
+        io.getOverviewSize(1, 1, &xsize, &ysize);
+        if( (xsize != 25) || (ysize != 50))
+        {
+            std::cout << "wrong overview size2" << std::endl;
+            return 1;
+        }
+        
+        std::cout << "Reading some overview data" << std::endl;
+        uint8_t *pReadOvData = (uint8_t*)calloc(OV_XSIZE * OV_YSIZE, sizeof(uint8_t));
+        io.readFromOverview(1, 0, pReadOvData, 0, 0, OV_XSIZE, OV_YSIZE, OV_XSIZE, OV_YSIZE, kealib::kea_8uint);
+        std::cout << "Read some overview data" << std::endl;
+
+        std::cout << "Comparing overview written and read data" << std::endl;
+        uint8_t *pTestOvData = createDataForType<uint8_t>(OV_XSIZE, OV_YSIZE);
+        if( !compareData<uint8_t>(pReadOvData, pTestOvData, OV_XSIZE, OV_YSIZE))
+        {
+            return 1;
+        }
+        free(pTestOvData);
+        free(pReadOvData);
+        std::cout << "Overview compared" << std::endl;
+        
+        io.close();
+    }
+    catch(const kealib::KEAException &e)
+    {
+        fprintf(stderr, "Exception raised: %s\n", e.what());
+        return 1;
+    }
+    printf("Success\n");
+
+    return 0;
+}
